Move timing tests out
